requirements.txt
pymupdf
sentence-transformers
faiss-cpu
torch
transformers
bitsandbytes
numpy
fastapi
uvicorn[standard]
python-multipart
reportlab
python-docx
